+++ files/gitlab-cookbooks/gitlab/libraries/helpers/mattermost_helper.rb
@@ -1,4 +1,3 @@
-require_relative 'shell_out_helper'
 require_relative 'authorizer_helper'
 
 class MattermostHelper