+++ files/gitlab-cookbooks/gitlab/libraries/matchers.rb
@@ -1,5 +1,0 @@
-if defined?(ChefSpec)
-  def create_templatesymlink(message)
-    ChefSpec::Matchers::ResourceMatcher.new(:templatesymlink, :create, message)
-  end
-end