+++ files/gitlab-cookbooks/gitlab/metadata.rb
@@ -9,6 +9,5 @@ recipe "gitlab", "Configures GitLab from Omnibus"
 
 supports "ubuntu"
 
-depends "runit"
-depends "package"
+depends "omnibus"
 depends 'registry'