+++ files/gitlab-cookbooks/omnibus/libraries/matchers.rb
@@ -1,0 +1,12 @@
+if defined?(ChefSpec)
+  def create_templatesymlink(message)
+    ChefSpec::Matchers::ResourceMatcher.new(:templatesymlink, :create, message)
+  end
+
+  # postgresql_database custom resource matchers
+  ChefSpec.define_matcher :postgresql_database
+
+  def create_postgresql_database(owner)
+    ChefSpec::Matchers::ResourceMatcher.new(:postgresql_database, :create, owner)
+  end
+end