+++ files/gitlab-cookbooks/repmgr/attributes/external.rb
@@ -1,0 +1,1 @@
+default['gitlab']['postgresql']['custom_pg_hba_entries']['repmgr'] = []