+++ lib/gitlab/package_repository.rb
@@ -6,8 +6,6 @@ class PackageRepository
     return ENV['PACKAGECLOUD_REPO'] if ENV['PACKAGECLOUD_REPO'] && !ENV['PACKAGECLOUD_REPO'].empty?
     # Repository for raspberry pi
     return ENV['RASPBERRY_REPO'] if ENV['RASPBERRY_REPO'] && !ENV['RASPBERRY_REPO'].empty?
-    # Repository for nightly build
-    return ENV['NIGHTLY_REPO'] if ENV['NIGHTLY_REPO'] && !ENV['NIGHTLY_REPO'].empty?
 
     rc_repository = repository_for_rc
     if rc_repository
@@ -71,7 +69,7 @@ class PackageRepository
       # to Scientific and Oracle Linux repositories
       if platform.start_with?("el/")
         %w(scientific ol).each do |distro|
-          platform_path = platform.tr('el', distro)
+          platform_path = platform.gsub('el', distro)
 
           list << "#{target_repository}/#{platform_path} #{package_path}"
         end