+++ spec/gitlab/package_repository_spec.rb
@@ -41,18 +41,6 @@ describe PackageRepository do
       end
     end
 
-    shared_examples 'with a nightly repository' do
-      context 'with nightly repo' do
-        before do
-          set_nightly_env_variable
-        end
-
-        it 'uses the nightly repository' do
-          expect(repo.target).to eq('nightly-builds')
-        end
-      end
-    end
-
     shared_examples 'with raspberry pi repo' do
       context 'with raspberry pi repo' do
         before do
@@ -75,7 +63,6 @@ describe PackageRepository do
       end
 
       it_behaves_like 'with an override repository'
-      it_behaves_like 'with a nightly repository'
       it_behaves_like 'with raspberry pi repo'
     end
 
@@ -94,7 +81,6 @@ describe PackageRepository do
         end
 
         it_behaves_like 'with an override repository'
-        it_behaves_like 'with a nightly repository'
         it_behaves_like 'with raspberry pi repo'
       end
 
@@ -108,7 +94,6 @@ describe PackageRepository do
         end
 
         it_behaves_like 'with an override repository'
-        it_behaves_like 'with a nightly repository'
         it_behaves_like 'with raspberry pi repo'
       end
     end
@@ -134,7 +119,7 @@ describe PackageRepository do
 
           it 'in dry run mode prints the upload commands' do
             expect { repo.upload('my-staging-repository', true) }.to output(%r{Uploading...\n}).to_stdout
-            expect { repo.upload('my-staging-repository', true) }.to output(%r{bin/package_cloud push gitlab/my-staging-repository/sc/6 pkg/el-6/gitlab-ce.rpm --url=https://packages.gitlab.com\n}).to_stdout
+            expect { repo.upload('my-staging-repository', true) }.to output(%r{bin/package_cloud push gitlab/my-staging-repository/scientific/6 pkg/el-6/gitlab-ce.rpm --url=https://packages.gitlab.com\n}).to_stdout
             expect { repo.upload('my-staging-repository', true) }.to output(%r{bin/package_cloud push gitlab/my-staging-repository/ol/6 pkg/el-6/gitlab-ce.rpm --url=https://packages.gitlab.com\n}).to_stdout
             expect { repo.upload('my-staging-repository', true) }.to output(%r{bin/package_cloud push gitlab/my-staging-repository/el/6 pkg/el-6/gitlab-ce.rpm --url=https://packages.gitlab.com\n}).to_stdout
           end
@@ -163,13 +148,12 @@ describe PackageRepository do
           before do
             stub_env_var('PACKAGECLOUD_REPO', nil)
             stub_env_var('RASPBERRY_REPO', nil)
-            stub_env_var('NIGHTLY_REPO', nil)
             allow_any_instance_of(PackageRepository).to receive(:repository_for_rc).and_return(nil)
           end
 
           context 'of EE' do
             before do
-              stub_env_var('ee', 'true')
+              stub_is_ee(true)
             end
 
             it 'in dry run mode prints the upload commands' do
@@ -180,7 +164,7 @@ describe PackageRepository do
 
           context 'of CE' do
             before do
-              stub_env_var('ee', nil)
+              stub_is_ee(nil)
             end
 
             it 'in dry run mode prints the upload commands' do
@@ -197,8 +181,8 @@ describe PackageRepository do
           end
 
           it 'in dry run mode prints the upload commands' do
-            expect { repo.upload(nil, true) }.to output(%r{Uploading...\n}).to_stdout
-            expect { repo.upload(nil, true) }.to output(%r{bin/package_cloud push gitlab/nightly-builds/ubuntu/xenial pkg/ubuntu-xenial/gitlab.deb --url=https://packages.gitlab.com\n}).to_stdout
+            expect { repo.upload(ENV['STAGING_REPO'], true) }.to output(%r{Uploading...\n}).to_stdout
+            expect { repo.upload(ENV['STAGING_REPO'], true) }.to output(%r{bin/package_cloud push gitlab/nightly-builds/ubuntu/xenial pkg/ubuntu-xenial/gitlab.deb --url=https://packages.gitlab.com\n}).to_stdout
           end
         end
 
@@ -231,19 +215,17 @@ describe PackageRepository do
 
   def set_all_env_variables
     stub_env_var("PACKAGECLOUD_REPO", "super-stable-1234")
-    stub_env_var("NIGHTLY_REPO", "nightly-builds")
     stub_env_var("RASPBERRY_REPO", "raspi")
   end
 
   def set_nightly_env_variable
     stub_env_var("PACKAGECLOUD_REPO", "")
-    stub_env_var("NIGHTLY_REPO", "nightly-builds")
     stub_env_var("RASPBERRY_REPO", "")
+    stub_env_var("STAGING_REPO", "nightly-builds")
   end
 
   def set_raspi_env_variable
     stub_env_var("PACKAGECLOUD_REPO", "")
-    stub_env_var("NIGHTLY_REPO", "nightly-builds")
     stub_env_var("RASPBERRY_REPO", "raspi")
   end
 end